# perf: Paging function edge cases fix