--- a/apps/explorer/lib/explorer/chain/optimism/output_root.ex
+++ b/apps/explorer/lib/explorer/chain/optimism/output_root.ex
@@ -47,23 +47,25 @@ defmodule Explorer.Chain.Optimism.OutputRoot do
   def list(options \\ []) do
     paging_options = Keyword.get(options, :paging_options, @default_paging_options)
 
-    base_query =
-      from(r in __MODULE__,
-        order_by: [desc: r.l2_output_index],
-        select: r
-      )
+    case paging_options do
+      %PagingOptions{key: {0}} ->
+        []
 
-    base_query
-    |> page_output_roots(paging_options)
-    |> limit(^paging_options.page_size)
-    |> select_repo(options).all()
+      _ ->
+        base_query =
+          from(r in __MODULE__,
+            order_by: [desc: r.l2_output_index],
+            select: r
+          )
+
+        base_query
+        |> page_output_roots(paging_options)
+        |> limit(^paging_options.page_size)
+        |> select_repo(options).all()
+    end
   end
 
   defp page_output_roots(query, %PagingOptions{key: nil}), do: query
-
-  defp page_output_roots(query, %PagingOptions{key: {0}}) do
-    query
-  end
 
   defp page_output_roots(query, %PagingOptions{key: {index}}) do
     from(r in query, where: r.l2_output_index < ^index)