--- a/apps/explorer/lib/explorer/chain/polygon_edge/reader.ex
+++ b/apps/explorer/lib/explorer/chain/polygon_edge/reader.ex
@@ -17,27 +17,33 @@ defmodule Explorer.Chain.PolygonEdge.Reader do
   def deposits(options \\ []) do
     paging_options = Keyword.get(options, :paging_options, default_paging_options())
 
-    base_query =
-      from(
-        de in DepositExecute,
-        inner_join: d in Deposit,
-        on: d.msg_id == de.msg_id and not is_nil(d.l1_timestamp),
-        select: %{
-          msg_id: de.msg_id,
-          from: d.from,
-          to: d.to,
-          l1_transaction_hash: d.l1_transaction_hash,
-          l1_timestamp: d.l1_timestamp,
-          success: de.success,
-          l2_transaction_hash: de.l2_transaction_hash
-        },
-        order_by: [desc: de.msg_id]
-      )
+    case paging_options do
+      %PagingOptions{key: {0}} ->
+        []
 
-    base_query
-    |> page_deposits_or_withdrawals(paging_options)
-    |> limit(^paging_options.page_size)
-    |> select_repo(options).all()
+      _ ->
+        base_query =
+          from(
+            de in DepositExecute,
+            inner_join: d in Deposit,
+            on: d.msg_id == de.msg_id and not is_nil(d.l1_timestamp),
+            select: %{
+              msg_id: de.msg_id,
+              from: d.from,
+              to: d.to,
+              l1_transaction_hash: d.l1_transaction_hash,
+              l1_timestamp: d.l1_timestamp,
+              success: de.success,
+              l2_transaction_hash: de.l2_transaction_hash
+            },
+            order_by: [desc: de.msg_id]
+          )
+
+        base_query
+        |> page_deposits_or_withdrawals(paging_options)
+        |> limit(^paging_options.page_size)
+        |> select_repo(options).all()
+    end
   end
 
   @spec deposits_count(list()) :: term() | nil
@@ -56,30 +62,36 @@ defmodule Explorer.Chain.PolygonEdge.Reader do
   def withdrawals(options \\ []) do
     paging_options = Keyword.get(options, :paging_options, default_paging_options())
 
-    base_query =
-      from(
-        w in Withdrawal,
-        left_join: we in WithdrawalExit,
-        on: we.msg_id == w.msg_id,
-        left_join: b in Block,
-        on: b.number == w.l2_block_number and b.consensus == true,
-        select: %{
-          msg_id: w.msg_id,
-          from: w.from,
-          to: w.to,
-          l2_transaction_hash: w.l2_transaction_hash,
-          l2_timestamp: b.timestamp,
-          success: we.success,
-          l1_transaction_hash: we.l1_transaction_hash
-        },
-        where: not is_nil(w.from),
-        order_by: [desc: w.msg_id]
-      )
+    case paging_options do
+      %PagingOptions{key: 0} ->
+        []
 
-    base_query
-    |> page_deposits_or_withdrawals(paging_options)
-    |> limit(^paging_options.page_size)
-    |> select_repo(options).all()
+      _ ->
+        base_query =
+          from(
+            w in Withdrawal,
+            left_join: we in WithdrawalExit,
+            on: we.msg_id == w.msg_id,
+            left_join: b in Block,
+            on: b.number == w.l2_block_number and b.consensus == true,
+            select: %{
+              msg_id: w.msg_id,
+              from: w.from,
+              to: w.to,
+              l2_transaction_hash: w.l2_transaction_hash,
+              l2_timestamp: b.timestamp,
+              success: we.success,
+              l1_transaction_hash: we.l1_transaction_hash
+            },
+            where: not is_nil(w.from),
+            order_by: [desc: w.msg_id]
+          )
+
+        base_query
+        |> page_deposits_or_withdrawals(paging_options)
+        |> limit(^paging_options.page_size)
+        |> select_repo(options).all()
+    end
   end
 
   @spec withdrawals_count(list()) :: term() | nil