… apps/block_scout_web/test/block_scout_web/controllers/address_internal_transaction_controller_test.exs
… apps/explorer/lib/explorer/chain/shibarium/reader.ex
… apps/explorer/lib/explorer/chain/token/instance.ex
… apps/explorer/lib/explorer/chain/withdrawal.ex
… apps/explorer/lib/explorer/chain/zksync/reader.ex
… apps/explorer/test/explorer/chain/transaction_test.exs