Resolves https://github.com/blockscout/blockscout/issues/9819

## Motivation

Skipping of edge cases in pagination leads to heavy meaningless queries being sent to the DB.

## Changelog

If 0 is sent for some of pagination key, we should avoid constructing query where `key < 0` since the DB doesn't store negative values for indexes, values or block numbers.

For instance, execution plan of some of the queries with not sanitized query:

```
eth_sepolia=> explain analyze SELECT t0.*, (SELECT transaction_hash FROM token_transfers WHERE transaction_hash = t0.hash LIMIT 1) IS NOT NULL FROM transactions AS t0 WHERE ((t0.block_number < 5613705) OR ((t0.block_number = 5613705) AND (t0.index < 0))) AND (NOT (t0.block_number IS NULL) AND NOT (t0.index IS NULL)) ORDER BY t0.block_number DESC, t0.index DESC LIMIT 51;
                                                                                             QUERY PLAN
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.57..232.54 rows=51 width=601) (actual time=83365.545..83375.112 rows=51 loops=1)
   ->  Index Scan using transactions_recent_collated_index on transactions t0  (cost=0.57..846752412.92 rows=186163165 width=601) (actual time=83365.545..83375.108 rows=51 loops=1)
         Index Cond: ((block_number IS NOT NULL) AND (index IS NOT NULL))
         Filter: ((block_number < 5613705) OR ((block_number = 5613705) AND (index < 0)))
         Rows Removed by Filter: 8215939
         SubPlan 1
           ->  Limit  (cost=0.70..0.97 rows=1 width=33) (actual time=0.190..0.190 rows=0 loops=51)
                 ->  Index Only Scan using token_transfers_transaction_hash_log_index_index on token_transfers  (cost=0.70..2900.16 rows=10730 width=33) (actual time=0.190..0.190 rows=0 loops=51)
                       Index Cond: (transaction_hash = t0.hash)
                       Heap Fetches: 9
 Planning Time: 0.239 ms
 Execution Time: 83375.138 ms
```

After sanitization:

```
eth_sepolia=> explain analyze SELECT t0.*, (SELECT transaction_hash FROM token_transfers WHERE transaction_hash = t0.hash LIMIT 1) IS NOT NULL FROM transactions AS t0 WHERE ((t0.block_number < 5613705)) AND (NOT (t0.block_number IS NULL) AND NOT (t0.index IS NULL)) ORDER BY t0.block_number DESC, t0.index DESC LIMIT 51;
                                                                                             QUERY PLAN
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.57..224.77 rows=51 width=601) (actual time=0.982..21.060 rows=51 loops=1)
   ->  Index Scan using transactions_recent_collated_index on transactions t0  (cost=0.57..818372415.96 rows=186163165 width=601) (actual time=0.981..21.056 rows=51 loops=1)
         Index Cond: ((block_number < 5613705) AND (block_number IS NOT NULL) AND (index IS NOT NULL))
         SubPlan 1
           ->  Limit  (cost=0.70..0.97 rows=1 width=33) (actual time=0.361..0.361 rows=0 loops=51)
                 ->  Index Only Scan using token_transfers_transaction_hash_log_index_index on token_transfers  (cost=0.70..2900.16 rows=10730 width=33) (actual time=0.361..0.361 rows=0 loops=51)
                       Index Cond: (transaction_hash = t0.hash)
                       Heap Fetches: 9
 Planning Time: 0.249 ms
 Execution Time: 21.105 ms
(10 rows)
```

## Checklist for your Pull Request (PR)

  - [ ] If I added new functionality, I added tests covering it.
  - [ ] If I fixed a bug, I added a regression test to prevent the bug from silently reappearing again.
  - [ ] I checked whether I should update the docs and did so by submitting a PR to https://github.com/blockscout/docs
  - [ ] If I added/changed/removed ENV var, I submitted a PR to https://github.com/blockscout/docs to update the list of env vars at https://github.com/blockscout/docs/blob/master/for-developers/information-and-settings/env-variables.md and I updated the version to `master` in the Version column. Changes will be reflected in this table: https://docs.blockscout.com/for-developers/information-and-settings/env-variables. 
  - [ ] If I added new DB indices, I checked, that they are not redundant with PGHero or other tools.
  - [ ] If I added/removed chain type, I modified the Github CI matrix and PR labels accordingly.