@k1rill-fedoseev 2024-04-15
Looks like we have a non-optimal keyset pagination in its current form. Here are some queries I've tried:
```sql
EXPLAIN ANALYSE
SELECT *
FROM transactions
WHERE block_number < 5506536
   OR block_number = 5506536 AND index < 0
ORDER BY transactions.block_number DESC, index DESC
LIMIT 50

+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
|QUERY PLAN                                                                                                                                                                        |
+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
|Limit  (cost=0.57..190.10 rows=50 width=600) (actual time=261711.299..261712.260 rows=50 loops=1)                                                                                 |
|  ->  Index Scan using transactions_recent_collated_index on transactions  (cost=0.57..669875126.15 rows=176722081 width=600) (actual time=261711.298..261712.257 rows=50 loops=1)|
|        Filter: ((block_number < 5506536) OR ((block_number = 5506536) AND (index < 0)))                                                                                          |
|        Rows Removed by Filter: 19515356                                                                                                                                          |
|Planning Time: 0.157 ms                                                                                                                                                           |
|Execution Time: 261712.282 ms                                                                                                                                                     |
+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+


EXPLAIN ANALYSE
SELECT *
FROM transactions
WHERE (block_number, index) < (5506536, 0)
ORDER BY transactions.block_number DESC, index DESC
LIMIT 50

+------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
|QUERY PLAN                                                                                                                                                              |
+------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
|Limit  (cost=0.57..171.73 rows=50 width=600) (actual time=0.375..1.608 rows=50 loops=1)                                                                                 |
|  ->  Index Scan using transactions_recent_collated_index on transactions  (cost=0.57..604938488.73 rows=176722081 width=600) (actual time=0.374..1.604 rows=50 loops=1)|
|        Index Cond: (ROW(block_number, index) < ROW(5506536, 0))                                                                                                        |
|Planning Time: 0.136 ms                                                                                                                                                 |
|Execution Time: 1.630 ms                                                                                                                                                |
+------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
```

It looks like that the problem is that postgres doesn't use proper index scan with optimal index cond when we use pagination in the form of `col_A < X OR col_A = X AND col_B < Y`. I think we may need to refactor pagination clauses to use format like `(col_A, col_B) < (X, Y)` whenever possible.